+++ CHANGELOG.md
@@ -36,6 +36,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 
+- Fix a logic issue making MM access way too restrictive in some cases.
 - Fix traps sometimes triggering at unsafe times.
 - Fixed MM Hyper enemy crash during Majora boss fight phase 3 when he spawns tops.
 - Logic now accounts for Bronze Scale or Iron Boots when coming from Gerudo Valley in Lake Hylia.